Fixed method's repeated DB connections Fixed query exception thrown when ResultSet was unused Bettered SQL methods with error checking Moved Connection conn to class attribute Successfully returned data with testQuery()

+++ app/src/main/java/com/team3/DBService.java
@@ -22,14 +22,15 @@ public class DBService {
     private final String url = "jdbc:postgresql://ec2-35-175-68-90.compute-1.amazonaws.com:5432/da1l74klnmbaus?password=d0c5b2a8fb21acb46bd2f2046696c971d61b24a42206531ddfcb6552e5eb4e76&sslmode=require&user=ujdocfowronvnp";
     private final String user = "ujdocfowronvnp";
     private final String password = "d0c5b2a8fb21acb46bd2f2046696c971d61b24a42206531ddfcb6552e5eb4e76";
+    public static Connection conn;
 
     DBService() {
-
+        conn = null;
     }
 
     private Connection getConnection() {
        
-        Connection conn = null;
+        conn = null;
         try {
             conn = DriverManager.getConnection(url, user, password);
             System.out.println("Connected to the PostgreSQL server successfully.");
@@ -44,16 +45,38 @@ public class DBService {
         getConnection();
     }
 
+    // Test returning query results to print in the terminal
+    public void testQuery() {
+        try {
+            Statement stmt = conn.createStatement();
+            
+            ResultSet rs = stmt.executeQuery("SELECT first_name, last_name FROM player");
+
+            while (rs.next()) {
+                System.out.print(rs.getString("first_name") + " ");
+                System.out.println(rs.getString("last_name"));
+            }
+        } catch (SQLException e) {
+            System.out.println("Failed to execute testQuery\n");
+            System.out.println(e.getMessage());
+        }
+    }
+
     //Creating a player in the database
     //Use SQL commands to modifiy the database with these methods
     public void create_player(int id, String first_name, String last_name, String codename){
         try {
-        Statement st = getConnection().createStatement();
+            if (!conn.isClosed()) {
+                Statement st = conn.createStatement();
 
-        // Revisit later and use PreparedStatement instead. Safer and more modular
-        String query = "INSERT INTO player VALUES ('" + id + "', '" + first_name + "', '" + last_name + "', '" + codename + "');";
-        ResultSet rs = st.executeQuery(query);
+                // Revisit later and use PreparedStatement instead. Safer and more modular
+                String create_command = "INSERT INTO player VALUES ('" + id + "', '" + first_name + "', '" + last_name + "', '" + codename + "');";
+                st.executeUpdate(create_command);
+            } else {
+                System.out.println("Can't execute create_player without a connection.\n");
+            }
         } catch (SQLException e) {
+            System.out.println("Failed to execute create_player\n");
             System.out.println(e.getMessage());
         }
     }
@@ -61,11 +84,15 @@ public class DBService {
     //Removing a player in the Database
     public void remove_player(int id){
         try {
-            Statement st = getConnection().createStatement();
-            
-            // Revisit later and use PreparedStatement instead. Safer and more modular
-            String query = "DELETE FROM player WHERE id = '" + id + "';";
-            ResultSet rs = st.executeQuery(query);
+            if (!conn.isClosed()) {
+                Statement st = conn.createStatement();
+                
+                // Revisit later and use PreparedStatement instead. Safer and more modular
+                String remove_command = "DELETE FROM player WHERE id = '" + id + "';";
+                st.executeUpdate(remove_command);
+            } else {
+                System.out.println("Can't execute remove_player without a connection.\n");
+            }
         } catch (SQLException e) {
             System.out.println(e.getMessage());
         }
@@ -74,11 +101,15 @@ public class DBService {
     //Modifing a player in the Database
     public void modify_player(int id, String first_name, String last_name, String codename){
         try {
-            Statement st = getConnection().createStatement();
-            
-            // Revisit later and use PreparedStatement instead. Safer and more modular
-            String query = "UPDATE player SET first_name = '" + first_name + "', last_name = '" + last_name + "', codename = '" + codename + "'  WHERE id = '" + id + "';";
-            ResultSet rs = st.executeQuery(query);
+            if (!conn.isClosed()) {
+                Statement st = conn.createStatement();
+                
+                // Revisit later and use PreparedStatement instead. Safer and more modular
+                String modify_command = "UPDATE player SET first_name = '" + first_name + "', last_name = '" + last_name + "', codename = '" + codename + "'  WHERE id = '" + id + "';";
+                st.executeUpdate(modify_command);
+            } else {
+                System.out.println("Can't execute modify_player without a connection.\n");
+            }
         } catch (SQLException e) {
             System.out.println(e.getMessage());
         }
